+++ poly-benches/benches/fft.rs
@@ -1,5 +1,3 @@
-use rand;
-
 extern crate criterion;
 
 use ark_ff::FftField;
@@ -41,7 +39,7 @@ fn setup_bench(c: &mut Criterion, name: &str, bench_fn: fn(&mut Bencher, &usize)
 }
 
 fn fft_common_setup<F: FftField, D: EvaluationDomain<F>>(degree: usize) -> (D, Vec<F>) {
-    let mut rng = &mut rand::thread_rng();
+    let mut rng = &mut ark_std::test_rng();
     let domain = D::new(degree).unwrap();
     let a = DensePolynomial::<F>::rand(degree - 1, &mut rng)
         .coeffs()