+++ poly/src/domain/general.rs
@@ -263,10 +263,10 @@ mod tests {
     use crate::polynomial::Polynomial;
     use crate::{EvaluationDomain, GeneralEvaluationDomain};
     use ark_ff::Zero;
+    use ark_std::rand::Rng;
     use ark_std::test_rng;
     use ark_test_curves::bls12_381::Fr;
     use ark_test_curves::bn384_small_two_adicity::Fr as BNFr;
-    use rand::Rng;
 
     #[test]
     fn vanishing_polynomial_evaluation() {