+++ poly/src/domain/radix2/mod.rs
@@ -214,9 +214,9 @@ mod tests {
     use crate::polynomial::{univariate::*, Polynomial, UVPolynomial};
     use crate::{EvaluationDomain, Radix2EvaluationDomain};
     use ark_ff::{FftField, Field, One, UniformRand, Zero};
+    use ark_std::rand::Rng;
     use ark_std::test_rng;
     use ark_test_curves::bls12_381::Fr;
-    use rand::Rng;
 
     #[test]
     fn vanishing_polynomial_evaluation() {