+++ poly/src/evaluations/multivariate/multilinear/dense.rs
@@ -6,9 +6,9 @@ use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, Read, Serializatio
 use ark_std::fmt;
 use ark_std::fmt::Formatter;
 use ark_std::ops::{Add, AddAssign, Index, Neg, Sub, SubAssign};
+use ark_std::rand::Rng;
 use ark_std::slice::{Iter, IterMut};
 use ark_std::vec::Vec;
-use rand::Rng;
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;