+++ test-templates/src/curves.rs
@@ -5,13 +5,11 @@ use ark_ec::{
 use ark_ff::{Field, One, PrimeField, UniformRand, Zero};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, SWFlags, SerializationError};
 use ark_std::{io::Cursor, vec::Vec};
-use rand::SeedableRng;
-use rand_xorshift::XorShiftRng;
 
 pub const ITERATIONS: usize = 10;
 
 fn random_addition_test<G: ProjectiveCurve>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = G::rand(&mut rng);
@@ -90,7 +88,7 @@ fn random_addition_test<G: ProjectiveCurve>() {
 }
 
 fn random_multiplication_test<G: ProjectiveCurve>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let mut a = G::rand(&mut rng);
@@ -122,7 +120,7 @@ fn random_multiplication_test<G: ProjectiveCurve>() {
 }
 
 fn random_doubling_test<G: ProjectiveCurve>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let mut a = G::rand(&mut rng);
@@ -149,7 +147,7 @@ fn random_doubling_test<G: ProjectiveCurve>() {
 }
 
 fn random_negation_test<G: ProjectiveCurve>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let r = G::rand(&mut rng);
@@ -178,7 +176,7 @@ fn random_negation_test<G: ProjectiveCurve>() {
 }
 
 fn random_transformation_test<G: ProjectiveCurve>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let g = G::rand(&mut rng);
@@ -193,7 +191,7 @@ fn random_transformation_test<G: ProjectiveCurve>() {
             .map(|_| G::rand(&mut rng).double())
             .collect::<Vec<_>>();
 
-        use rand::distributions::{Distribution, Uniform};
+        use ark_std::rand::distributions::{Distribution, Uniform};
         let between = Uniform::from(0..ITERATIONS);
         // Sprinkle in some normalized points
         for _ in 0..5 {
@@ -219,7 +217,7 @@ fn random_transformation_test<G: ProjectiveCurve>() {
 }
 
 pub fn curve_tests<G: ProjectiveCurve>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     // Negation edge case with zero.
     {
@@ -296,7 +294,7 @@ pub fn sw_from_random_bytes<P: SWModelParameters>() {
 
     let buf_size = GroupAffine::<P>::zero().serialized_size();
 
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = GroupProjective::<P>::rand(&mut rng);
@@ -319,7 +317,7 @@ pub fn sw_curve_serialization_test<P: SWModelParameters>() {
 
     let buf_size = GroupAffine::<P>::zero().serialized_size();
 
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = GroupProjective::<P>::rand(&mut rng);
@@ -431,7 +429,7 @@ where
 
     let buf_size = GroupAffine::<P>::zero().serialized_size();
 
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = GroupProjective::<P>::rand(&mut rng);
@@ -466,7 +464,7 @@ pub fn edwards_curve_serialization_test<P: TEModelParameters>() {
 
     let buf_size = GroupAffine::<P>::zero().serialized_size();
 
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = GroupProjective::<P>::rand(&mut rng);