+++ test-templates/src/fields.rs
@@ -2,8 +2,7 @@
 use ark_ff::fields::{FftField, FftParameters, Field, LegendreSymbol, PrimeField, SquareRootField};
 use ark_serialize::{buffer_bit_byte_size, Flags, SWFlags};
 use ark_std::io::Cursor;
-use rand::{Rng, SeedableRng};
-use rand_xorshift::XorShiftRng;
+use ark_std::rand::Rng;
 
 pub const ITERATIONS: u32 = 40;
 
@@ -154,7 +153,7 @@ fn random_expansion_tests<F: Field, R: Rng>(rng: &mut R) {
 }
 
 fn random_field_tests<F: Field>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     random_negation_tests::<F, _>(&mut rng);
     random_addition_tests::<F, _>(&mut rng);
@@ -189,7 +188,7 @@ fn random_field_tests<F: Field>() {
 }
 
 fn random_sqrt_tests<F: SquareRootField>() {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = F::rand(&mut rng);
@@ -219,7 +218,7 @@ fn random_sqrt_tests<F: SquareRootField>() {
 
 pub fn from_str_test<F: PrimeField>() {
     {
-        let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+        let mut rng = ark_std::test_rng();
 
         for _ in 0..ITERATIONS {
             let n: u64 = rng.gen();
@@ -357,7 +356,7 @@ pub fn sqrt_field_test<F: SquareRootField>(elem: F) {
 }
 
 pub fn frobenius_test<F: Field, C: AsRef<[u64]>>(characteristic: C, maxpower: usize) {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = F::rand(&mut rng);
@@ -378,7 +377,7 @@ pub fn frobenius_test<F: Field, C: AsRef<[u64]>>(characteristic: C, maxpower: us
 }
 
 pub fn field_serialization_test<F: Field>(buf_size: usize) {
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
+    let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let a = F::rand(&mut rng);